head class added, some bugs are fixed.
I add head class and some essential html code.
And some bugs are fixed.

MINI on 2018-06-01-23:12(KST).
