--- Mutwo/Mutwo.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26228.4
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Mutwo", "Mutwo\Mutwo.vcxproj", "{81D7A965-D868-49B3-8C68-D082240C75DF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Debug|x64.ActiveCfg = Debug|x64
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Debug|x64.Build.0 = Debug|x64
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Debug|x86.ActiveCfg = Debug|Win32
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Debug|x86.Build.0 = Debug|Win32
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Release|x64.ActiveCfg = Release|x64
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Release|x64.Build.0 = Release|x64
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Release|x86.ActiveCfg = Release|Win32
-		{81D7A965-D868-49B3-8C68-D082240C75DF}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
