--- Mutwo/Mutwo/Debug/Mutwo.tlog/Mutwo.lastbuildstate
@@ -1,2 +1,0 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.14393.0
-Debug|Win32|C:\Users\정진우\Desktop\동고비\수업\오픈소스\팀프로젝트\MuTwo\Mutwo\|
