+++ Mutwo/Mutwo_ver1.0.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2027
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Mutwo_ver1.0", "Mutwo_ver1.0\Mutwo_ver1.0.vcxproj", "{878E860F-CDF0-4C53-AC66-B0196008E0F7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Debug|x64.ActiveCfg = Debug|x64
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Debug|x64.Build.0 = Debug|x64
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Debug|x86.ActiveCfg = Debug|Win32
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Debug|x86.Build.0 = Debug|Win32
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Release|x64.ActiveCfg = Release|x64
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Release|x64.Build.0 = Release|x64
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Release|x86.ActiveCfg = Release|Win32
+		{878E860F-CDF0-4C53-AC66-B0196008E0F7}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D2A8E429-5329-4E52-ACD2-81228BBAA844}
+	EndGlobalSection
+EndGlobal
