+++ Mutwo/Mutwo_ver1.0/Debug/Mutwo_ver1.0.tlog/Mutwo_ver1.0.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.16299.0
+Debug|Win32|C:\Users\win10se\source\repos\Mutwo_ver1.0\|
